Frontend 'Create Stock' Form Validation

+++ backend/DTOs/StockDTO.cs
@@ -1,7 +1,4 @@
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace backend.DTOs
 {
@@ -10,7 +7,7 @@ namespace backend.DTOs
         public Guid id { get; set; }
         public string code { get; set; }
         public string name { get; set; }
-        public int units { get; set; }
+        public float units { get; set; }
         public float purchase_price { get; set; }
         public string date_purchased { get; set; }
     }
